prepares server version
