+++ css/dash.css
@@ -1,5 +1,5 @@
 /**
- * part of https://github.com/rc-dukes/dash
+ * part of https://github.com/rc-dukes/dash fork of https://github.com/mattbradley/dash
  */
 html { overflow: hidden !important; }
 body { margin: 0; }
