+++ index.html
@@ -1,4 +1,4 @@
-<!-- Part of https://github.com/rc-dukes/dash/ forked of Dash -->
+<!-- part of https://github.com/rc-dukes/dash fork of https://github.com/mattbradley/dash -->
 <html>
 	<head>
         <title>Dash - WebGL Self-Driving Car Simulator</title>
