+++ package.json
@@ -9,6 +9,7 @@
   "author": "Matt Bradley",
   "license": "MIT",
   "devDependencies": {
+    "@types/node": "^13.1.6",
     "script-loader": "^0.7.2",
     "webpack": "^4.41.5",
     "webpack-cli": "^3.3.10",
